+++ src/utils.hpp
@@ -10,7 +10,7 @@ int min_indx_cpp(int k, int l);
 
 int max_indx_cpp(int k, int l);
 
-Eigen::VectorXd which_min_upper(Eigen::MatrixXd mat);
+Eigen::VectorXi which_min_upper(Eigen::MatrixXd mat);
 
 Eigen::MatrixXd inverse(Eigen::MatrixXd A);
